Move test dependencies
Move test dependencies into 'test' poetry group, and modify the Docker build
to not install test dependencies.

Fixes #274
