--- libsys_airflow/plugins/tests/test_login.py
@@ -1,5 +1,0 @@
-from plugins.folio.login import folio_login
-
-
-def test_folio_login():
-    assert folio_login
