+++ pyproject.toml
@@ -15,17 +15,19 @@ pandas = "^1.4.1"
 pydantic = "^1.9.0"
 pymarc = "^4.1.3"
 folio-migration-tools = "1.7.8"
-pytest = "^7.1.1"
-pytest-mock = "^3.7.0"
+requests = "^2.28.2"
+pyyaml = "^6.0"
 
-[tool.poetry.dev-dependencies]
-apache-airflow = "2.5.3"
-apache-airflow-providers-postgres = "5.4.0"
+[tool.poetry.group.test.dependencies]
 coveralls = "~3.3.1"
 coverage = "~6.5.0"
 flake8 = "^6.0.0"
-folioclient = "^0.43.0"
-folio-uuid = "^0.2.8"
+pytest = "^7.1.1"
+pytest-mock = "^3.7.0"
+# these are needed for running tests outside of the airflow docker container
+# the airflow version here should be kept in sync with the docker image
+apache-airflow = "2.5.3"
+apache-airflow-providers-postgres = "5.4.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
