+++ tests/test_login.py
@@ -1,0 +1,5 @@
+from libsys_airflow.plugins.folio.login import folio_login
+
+
+def test_folio_login():
+    assert folio_login
